# Too strict requirements on u-turns