--- a/src/baldr/graphtile.cc
+++ b/src/baldr/graphtile.cc
@@ -548,7 +548,8 @@ iterable_t<const DirectedEdge> GraphTile::GetDirectedEdges(const GraphId& node) 
     return GetDirectedEdges(nodeinfo);
   }
   throw std::runtime_error(
-      "GraphTile NodeInfo index out of bounds 4: " + std::to_string(node.tileid()) + "," +
+      std::string(__FILE__) + ":" + std::to_string(__LINE__) +
+      " GraphTile NodeInfo index out of bounds: " + std::to_string(node.tileid()) + "," +
       std::to_string(node.level()) + "," + std::to_string(node.id()) +
       " nodecount= " + std::to_string(header_->nodecount()));
 }
@@ -560,7 +561,8 @@ iterable_t<const DirectedEdge> GraphTile::GetDirectedEdges(const size_t idx) con
     return iterable_t<const DirectedEdge>{edge, nodeinfo.edge_count()};
   }
   throw std::runtime_error(
-      "GraphTile NodeInfo index out of bounds 5: " + std::to_string(header_->graphid().tileid()) +
+      std::string(__FILE__) + ":" + std::to_string(__LINE__) +
+      " GraphTile NodeInfo index out of bounds 5: " + std::to_string(header_->graphid().tileid()) +
       "," + std::to_string(header_->graphid().level()) + "," + std::to_string(idx) +
       " nodecount= " + std::to_string(header_->nodecount()));
 }