--- a/valhalla/baldr/graphtile.h
+++ b/valhalla/baldr/graphtile.h
@@ -135,7 +135,7 @@ public:
       return &nodes_[node.id()];
     }
     throw std::runtime_error(
-        "GraphTile NodeInfo index out of bounds: " + std::to_string(node.tileid()) + "," +
+        "GraphTile NodeInfo index out of bounds 1: " + std::to_string(node.tileid()) + "," +
         std::to_string(node.level()) + "," + std::to_string(node.id()) +
         " nodecount= " + std::to_string(header_->nodecount()));
   }
@@ -150,7 +150,7 @@ public:
       return &nodes_[idx];
     }
     throw std::runtime_error(
-        "GraphTile NodeInfo index out of bounds: " + std::to_string(header_->graphid().tileid()) +
+        "GraphTile NodeInfo index out of bounds 2: " + std::to_string(header_->graphid().tileid()) +
         "," + std::to_string(header_->graphid().level()) + "," + std::to_string(idx) +
         " nodecount= " + std::to_string(header_->nodecount()));
   }
@@ -263,7 +263,7 @@ public:
       return GetNodeTransitions(nodeinfo);
     }
     throw std::runtime_error(
-        "GraphTile NodeInfo index out of bounds: " + std::to_string(node.tileid()) + "," +
+        "GraphTile NodeInfo index out of bounds 3: " + std::to_string(node.tileid()) + "," +
         std::to_string(node.level()) + "," + std::to_string(node.id()) +
         " nodecount= " + std::to_string(header_->nodecount()));
   }