--- a/valhalla/thor/bidirectional_astar.h
+++ b/valhalla/thor/bidirectional_astar.h
@@ -35,6 +35,8 @@ struct CandidateConnection {
   }
 };
 
+struct EdgeMetadata;
+
 /**
  * Bidirectional A* algorithm. Method for finding least-cost path.
  */
@@ -125,20 +127,37 @@ protected:
    */
   void ExpandForward(baldr::GraphReader& graphreader,
                      const baldr::GraphId& node,
-                     const sif::BDEdgeLabel& pred,
+                     sif::BDEdgeLabel& pred,
                      const uint32_t pred_idx,
                      const bool from_transition);
+  // Private helper function for `ExpandForward`
+  bool ExpandForwardInner(baldr::GraphReader& graphreader,
+                          const sif::BDEdgeLabel& pred,
+                          const baldr::NodeInfo* nodeinfo,
+                          const uint32_t pred_idx,
+                          const EdgeMetadata& meta,
+                          uint32_t& shortcuts,
+                          const baldr::GraphTile* tile);
 
   /**
    * Expand from the node along the reverse search path.
    */
   void ExpandReverse(baldr::GraphReader& graphreader,
                      const baldr::GraphId& node,
-                     const sif::BDEdgeLabel& pred,
+                     sif::BDEdgeLabel& pred,
                      const uint32_t pred_idx,
                      const baldr::DirectedEdge* opp_pred_edge,
                      const bool from_transition);
 
+  // Private helper function for `ExpandReverse`
+  bool ExpandReverseInner(baldr::GraphReader& graphreader,
+                          const sif::BDEdgeLabel& pred,
+                          const baldr::DirectedEdge* opp_pred_edge,
+                          const baldr::NodeInfo* nodeinfo,
+                          const uint32_t pred_idx,
+                          const EdgeMetadata& meta,
+                          uint32_t& shortcuts,
+                          const baldr::GraphTile* tile);
   /**
    * Add edges at the origin to the forward adjacency list.
    * @param  graphreader  Graph tile reader.