… CHANGELOG.md
… docs/api/index.md
… test/CMakeLists.txt
… test/data/whitelion_bristol_uk.osm.pbf