# Issue

 feat(thor): Adds runtime check of whether an edge is a deadend

This allows for time dependence of deadend-detection as well as resolving an issue
where a path only is a deadend in the reverse expansion of
bidirectional_astar.

Fixes #1982

## Performance hit

The bidirectional a-star takes a performance hit from this change. This is due to two reasons:

 1) Evaluating whether something is a deadend at runtime rather than using the pre-computed deadend bit. _We managed to write this runtime check inside the existing loop which we believe make this check basically a no-change to performance compared to the naive check._
 2) Actually relaxing the requirements for a u-turn means that more nodes will be expanded in the search.

I ran a few benchmarks on the common cross-country trip. I split the PR into two major commits. The first one simply refactors the `Expand*` functions and puts the loop contents into `Expand*Inner` functions to improve clarity of code. Logic should still be same as in `master`.

The second commit actually switches to the runtime detection of deadends. This is just a few lines of change from the previous commit in order to avoid ascribing too much of the performance hit to the refactor:

Benchmarked using cross-country route:
```fish
./valhalla_run_route \
  --config ~/data/valhalla.json \
  --multi-run 100 \
    -j '{"locations":[{"lat":32.697958,"lon":-117.174648, "street":"Park Boulevard"},{"lat":40.042553,"lon":-76.299115}],"costing":"auto"}' \
  | tee result \
  | grep average
```

```fish
New performance numbers

  derived_deadend (git:3c6439e51)

    #1 PathAlgorithm GetBestPath average: 196 ms
    #2 PathAlgorithm GetBestPath average: 201 ms

  Refactored Expand* functions but still same logic as in master (git:4432b4a8c)

    #1 PathAlgorithm GetBestPath average: 197 ms
    #2 PathAlgorithm GetBestPath average: 199 ms

  master (git:b14999693)

    #1 PathAlgorithm GetBestPath average: 183 ms
    #2 PathAlgorithm GetBestPath average: 184 ms

```

**Expanding more edges due to more deadends seem to carry a ~9% increase in wall-clock time on cross country route on my laptop**. 


## Open questions

If we go ahead with this change, the time-dependant a-star should probably be updated to use this logic as well.

## Tasklist

 - [] Add tests (Some added, investigating adding more)
 - [x] Review - you must request approval to merge any PR to master
 - [x] Add #fixes with the issue number that this PR addresses
 - [x] Generally use squash merge to rebase and clean comments before merging
 - [x] Update the [changelog](CHANGELOG.md)
 - [ ] Update relevant [documentation](docs/README.md)

## Requirements / Relations

 Link any requirements here. Other pull requests this PR is based on?