@dnesbitt61 2019-11-01
that is pretty serious performance reduction. Did you test with any pedestrian or bicycle routes to see what impact is seen there? Pedestrian basically allows Uturns anywhere so I assume there should be no change there. 
Did you look into stats on how large the increase in path iterations was (combined size of the edge labels - logged in FormPath)? Did you see roughly 50% more EdgeLabels?