@PureW 2019-11-01
> Did you look into stats on how large the increase in path iterations was (combined size of the edge labels - logged in FormPath)? Did you see roughly 50% more EdgeLabels?

@dnesbitt61 I updated the post above. Seems like roughly 70% more path-iterations are made with this change.

I've yet to investigate walking/bike routes. Working on those.