@dnesbitt61 2019-11-05
numbers look a lot better! I couldn't think of how a few more dead-end cases could cause such a dramatic increase.