@PureW 2019-11-06
Performance is hard. I've rerun benches this morning on my i7 desktop at home that doesn't do funny things like thermal throttling or adjusting schedulers depending on battery levels etc.

Inlinining the `ExpandXXXInner` functions _does_ seem to make a difference on that cpu/computer.I now measured 136ms vs 142ms, a 4% overhead with new code: https://github.com/valhalla/valhalla/pull/2018/commits/53b661c7df6281bd916bfdda9150a2abf5fcae6d

cc @dnesbitt61